Preparation of the project for the development of the functional
