+++ .gitignore
@@ -1,5 +1,5 @@
 .stack-work/
-hpars.cabal
+*.cabal
 *.swp
 *~
 .*
