--- ABin/start.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-
-../.stack-work/install/x86_64-linux-nopie/lts-12.1/8.4.3/bin/\
-text-tokenize-exe $@
-
-
-
