--- ADoc/Docs.txt
@@ -1,1 +1,0 @@
-https://github.com/hexresearch/text-tokenize/issues/1
