+++ package.yaml
@@ -1,6 +1,6 @@
 name:                text-tokenize
 version:             0.1.0.0
-github:              "githubuser/text-tokenize"
+github:              "hexresearch/text-tokenize"
 license:             BSD3
 author:              "Author name here"
 maintainer:          "example@example.com"
@@ -17,10 +17,11 @@ extra-source-files:
 # To avoid duplicated efforts in documentation and dealing with the
 # complications of embedding Haddock markup inside cabal files, it is
 # common to point users to the README.md file.
-description:         Please see the README on GitHub at <https://github.com/githubuser/text-tokenize#readme>
+description:         Please see the README on GitHub at <https://github.com/hexresearch/text-tokenize#readme>
 
 dependencies:
-- base >= 4.7 && < 5
+- base                  >= 4.7      && < 5
+- text                  >= 1.2.3.0  && < 1.3
 
 library:
   source-dirs: src
