--- text-tokenize.cabal
@@ -1,72 +1,0 @@
--- This file has been generated from package.yaml by hpack version 0.20.0.
---
--- see: https://github.com/sol/hpack
---
--- hash: 2dd80bfe2883cb591f2ce24c68d688033deaf4595ce4481272147e550d28f39b
-
-name:           text-tokenize
-version:        0.1.0.0
-description:    Please see the README on GitHub at <https://github.com/hexresearch/text-tokenize#readme>
-homepage:       https://github.com/hexresearch/text-tokenize#readme
-bug-reports:    https://github.com/hexresearch/text-tokenize/issues
-author:         Author name here
-maintainer:     example@example.com
-copyright:      2018 Author name here
-license:        BSD3
-license-file:   LICENSE
-build-type:     Simple
-cabal-version:  >= 1.10
-
-extra-source-files:
-    ChangeLog.md
-    README.md
-
-source-repository head
-  type: git
-  location: https://github.com/hexresearch/text-tokenize
-
-
-library
-  hs-source-dirs:
-      src
-  build-depends:
-      base                  >= 4.7      && < 5
-    , text                  >= 1.2.3.0  && < 1.3
-
-  exposed-modules:
-      TextTokenize.Parser
-    , TextTokenize.ParserTF
-
-  other-modules:
-      Paths_text_tokenize
-
-  default-language: Haskell2010
-
-
-executable text-tokenize-exe
-  main-is: Main.hs
-  hs-source-dirs:
-      app
-  ghc-options: -threaded -rtsopts -with-rtsopts=-N
-  build-depends:
-      base                  >= 4.7      && < 5
-    , text                  >= 1.2.3.0  && < 1.3
-    , text-tokenize
-
-  other-modules:
-      Paths_text_tokenize
-  default-language: Haskell2010
-
-
-test-suite text-tokenize-test
-  type: exitcode-stdio-1.0
-  main-is: Spec.hs
-  hs-source-dirs:
-      test
-  ghc-options: -threaded -rtsopts -with-rtsopts=-N
-  build-depends:
-      base >=4.7 && <5
-    , text-tokenize
-  other-modules:
-      Paths_text_tokenize
-  default-language: Haskell2010
